refactor: 리포지토리 위치 infra -> domain으로 변경
