+++ src/main/java/coffeemeet/server/chatting/history/domain/repository/ChattingMessageHistoryRepository.java
@@ -1,9 +1,9 @@
-package coffeemeet.server.chatting.history.infrastructure;
+package coffeemeet.server.chatting.history.domain.repository;
 
 import coffeemeet.server.chatting.history.domain.ChattingMessageHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChattingMessageHistoryRepository extends
-    JpaRepository<ChattingMessageHistory, Long> {
+    JpaRepository<ChattingMessageHistory, Long>, ChattingMessageHistoryJdbcRepository {
 
 }
