+++ src/main/java/coffeemeet/server/chatting/history/domain/repository/UserChattingHistoryRepository.java
@@ -1,4 +1,4 @@
-package coffeemeet.server.chatting.history.infrastructure;
+package coffeemeet.server.chatting.history.domain.repository;
 
 import coffeemeet.server.chatting.history.domain.ChattingRoomHistory;
 import coffeemeet.server.chatting.history.domain.UserChattingHistory;
